add TestTallyJailedValidator to simapp
