+++ x/gov/keeper/old_tally_test.go
@@ -7,47 +7,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/gov/types"
-	"github.com/cosmos/cosmos-sdk/x/staking"
 )
-
-func TestTallyJailedValidator(t *testing.T) {
-	ctx, _, _, keeper, sk, _ := createTestInput(t, false, 100)
-	createValidators(ctx, sk, []int64{25, 6, 7})
-
-	delTokens := sdk.TokensFromConsensusPower(10)
-	val2, found := sk.GetValidator(ctx, valOpAddr2)
-	require.True(t, found)
-	val3, found := sk.GetValidator(ctx, valOpAddr3)
-	require.True(t, found)
-
-	_, err := sk.Delegate(ctx, TestAddrs[0], delTokens, sdk.Unbonded, val2, true)
-	require.NoError(t, err)
-	_, err = sk.Delegate(ctx, TestAddrs[0], delTokens, sdk.Unbonded, val3, true)
-	require.NoError(t, err)
-
-	_ = staking.EndBlocker(ctx, sk)
-
-	sk.Jail(ctx, sdk.ConsAddress(val2.GetConsPubKey().Address()))
-
-	tp := TestProposal
-	proposal, err := keeper.SubmitProposal(ctx, tp)
-	require.NoError(t, err)
-	proposalID := proposal.ProposalID
-	proposal.Status = types.StatusVotingPeriod
-	keeper.SetProposal(ctx, proposal)
-
-	require.NoError(t, keeper.AddVote(ctx, proposalID, valAccAddr1, types.OptionYes))
-	require.NoError(t, keeper.AddVote(ctx, proposalID, valAccAddr2, types.OptionNo))
-	require.NoError(t, keeper.AddVote(ctx, proposalID, valAccAddr3, types.OptionNo))
-
-	proposal, ok := keeper.GetProposal(ctx, proposalID)
-	require.True(t, ok)
-	passes, burnDeposits, tallyResults := keeper.Tally(ctx, proposal)
-
-	require.True(t, passes)
-	require.False(t, burnDeposits)
-	require.False(t, tallyResults.Equals(types.EmptyTallyResult()))
-}
 
 func TestTallyValidatorMultipleDelegations(t *testing.T) {
 	ctx, _, _, keeper, sk, _ := createTestInput(t, false, 100)
